feat(bsw): add DET, SchM, CanTp modules + harden all 18 BSW modules
Phase 1 — DET (Default Error Tracer):
- New Det.h/Det.c: ring buffer error log, SIL_DIAG output, callback hook
- 123 Det_ReportError() calls added across all 18 BSW modules
- 14 unit tests (test_Det_asild.c)

Phase 2 — SchM (Schedule Manager):
- New SchM.h/SchM.c: critical sections (POSIX=no-op, STM32=IRQ disable)
- 5 modules protected: Com, Rte, Dem, WdgM, Can
- 11 unit tests (test_SchM_asild.c)

Phase 3 — CanTp (ISO 15765-2 transport protocol):
- New CanTp.h/CanTp.c: SF/FF/CF/FC state machine, reassembly, segmentation
- PduLengthType widened uint8→uint16 in ComStack_Types.h
- 25 unit tests (test_CanTp_asild.c)

Phase 4 — CanTp + Dcm integration:
- PduR: PDUR_DEST_CANTP routing, PduR_CanTpTransmit()
- Dcm: TX buffer 8→128, SecurityAccess (0x27), ECUReset (0x11)
- Dcm routes >7-byte responses through CanTp_Transmit
- 11 new Dcm tests (34 total), PduR CanTp mock added

Phase 5 — CVC ECU wiring:
- CVC main.c: CanTp_Init, CanTp_MainFunction (10ms), PduR diag→CanTp

22/22 test suites, 554 tests, 0 failures.

+++ firmware/shared/bsw/include/ComStack_Types.h
@@ -15,8 +15,8 @@
 /* PDU identifier type */
 typedef uint16 PduIdType;
 
-/* PDU length type */
-typedef uint8 PduLengthType;
+/* PDU length type (uint16 to support ISO-TP payloads >255 bytes) */
+typedef uint16 PduLengthType;
 
 /* PDU info structure -- passed through the COM stack */
 typedef struct {
